// dashboard/src/api/models/addressDetail.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { AddressDetailCountry } from "./addressDetailCountry";

export interface AddressDetail {
  /**
   * @maxLength 255
   * @nullable
   */
  line1?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  line2?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  locality?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  state?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  postal_code?: string | null;
  /** @nullable */
  country?: AddressDetailCountry;
  id: string;
  created_at: string;
  updated_at: string;
}